Module aws_sdk_pi::model

Modules

data_point

See DataPoint

dimension_group

See DimensionGroup

dimension_key_description

See DimensionKeyDescription

dimension_key_detail

See DimensionKeyDetail

metric_key_data_points

See MetricKeyDataPoints

metric_query

See MetricQuery

response_partition_key

See ResponsePartitionKey

response_resource_metric_key

See ResponseResourceMetricKey

Structs

DataPoint

A timestamp, and a single numerical value, which together represent a measurement at a particular point in time.

DimensionGroup

A logical grouping of Performance Insights metrics for a related subject area. For example, the

DimensionKeyDescription

An array of descriptions and aggregated values for

DimensionKeyDetail

An object that describes the details for a specified dimension.

MetricKeyDataPoints

A time-ordered series of data points, corresponding to a dimension of a Performance Insights

MetricQuery

A single query to be processed. You must provide the metric to query. If no other

ResponsePartitionKey

If PartitionBy was specified in a DescribeDimensionKeys

ResponseResourceMetricKey

An object describing a Performance Insights metric and one or more dimensions for that metric.

Enums

DetailStatus

ServiceType